Use a proper python package structure Python 3 version is kept. Python 2 removed.
